Remove wm_dialog_overlay
