+++ crates/term-wm-sys-ui-components/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod wm_bottom_panel;
 pub mod wm_debug_log;
-pub mod wm_dialog_overlay;
 pub mod wm_help_overlay;
 pub mod wm_keybinding_overlay;
 pub mod wm_menu_overlay;
@@ -8,7 +7,6 @@ pub mod wm_top_panel;
 
 pub use wm_bottom_panel::WmBottomPanelComponent;
 pub use wm_debug_log::{WmDebugLogComponent, install_panic_hook, set_global_debug_log};
-pub use wm_dialog_overlay::WmDialogOverlayComponent;
 pub use wm_help_overlay::WmHelpOverlayComponent;
 pub use wm_keybinding_overlay::WmKeybindingOverlayComponent;
 pub use wm_menu_overlay::WmMenuOverlay;
