+++ crates/term-wm-ui-components/src/lib.rs
@@ -5,6 +5,7 @@ pub mod list;
 pub mod markdown_viewer;
 pub mod menu;
 pub mod scroll_view;
+
 pub mod svg_image;
 pub mod terminal;
 pub mod text_renderer;
@@ -19,6 +20,7 @@ pub use menu::MenuComponent;
 pub use scroll_view::{
     ScrollViewComponent, ScrollbarAxis, ScrollbarDrag, render_scrollbar, render_scrollbar_oriented,
 };
+
 pub use svg_image::SvgImageComponent;
 pub use terminal::{TerminalComponent, default_shell, default_shell_command};
 pub use text_renderer::TextRendererComponent;
